+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## Unreleased
+
+* Add `-d` flag to make a "dry run": show the commits that would be deployed
+  without creating a deployment. (mrnugget)
+
 ## 1.0.0 - 09. September 2015
 
 * Initial release